1b_data-modeling-cassandra/requirements.txt
cassandra-driver==3.21.0